Added a "Trader" tile
Added a "Trader" tile to test for multi-tile functionality. previously the tile spawner was only able to spawn 1 kind of base tile which was stored by the tile spawner itself. Now each tile has a list of base tiles it 'can' spawn and has a method to choose it based on the 'percentage' chance each tile has. Next I want to implement a per tile spaawner class variable to indicate the maximum for the tile types (only want 1-2 traders per dungeon).